Symphony project/requirements.txt
# FastAPI and web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Free AI APIs
google-generativeai==0.3.2
huggingface-hub==0.20.1
transformers==4.36.0
torch==2.1.0
requests==2.31.0

# Database and cache
sqlalchemy==2.0.23
aiosqlite==0.19.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
colorama==0.4.6
aiofiles==23.2.1
celery==5.3.4